requirements.txt
Django==4.2.8
django-cors-headers==4.3.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
requests==2.31.0
